project - Copy (2)/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
kafka-python==2.0.2
requests==2.26.0
python-dateutil==2.8.2
tk==0.1.0 

pip install -r requirements.txt